+++ controllers/git_server.go
@@ -270,7 +270,7 @@ func (r *SFController) DeployGitServer() bool {
 
 	annotations := map[string]string{
 		"system-config": utils.Checksum([]byte(preInitScript)),
-		"image":         base.GitServerImage,
+		"image":         base.GitServerImage(),
 		"fqdn":          r.cr.Spec.FQDN,
 		"serial":        "2",
 	}
@@ -287,7 +287,7 @@ func (r *SFController) DeployGitServer() bool {
 	}
 
 	// Create the statefulset
-	sts := r.mkStatefulSet(GitServerIdent, base.GitServerImage, r.getStorageConfOrDefault(r.cr.Spec.GitServer.Storage), apiv1.ReadWriteOnce)
+	sts := r.mkStatefulSet(GitServerIdent, base.GitServerImage(), r.getStorageConfOrDefault(r.cr.Spec.GitServer.Storage), apiv1.ReadWriteOnce)
 	sts.Spec.Template.ObjectMeta.Annotations = annotations
 	GSVolumeMountsRO := []apiv1.VolumeMount{
 		{
@@ -300,7 +300,7 @@ func (r *SFController) DeployGitServer() bool {
 	sts.Spec.Template.Spec.Containers[0].Command = []string{"git", "daemon", "--base-path=/git", "--export-all"}
 
 	// Add a second container to serve the git-server with RW access (should not be exposed)
-	containerRW := base.MkContainer("git-daemon-rw", base.GitServerImage)
+	containerRW := base.MkContainer("git-daemon-rw", base.GitServerImage())
 	containerRW.Command = []string{"git", "daemon", "--base-path=/git",
 		"--enable=receive-pack", "--export-all", "--port=" + strconv.Itoa(gsGitPortRW)}
 	containerRW.VolumeMounts = []apiv1.VolumeMount{
@@ -320,7 +320,7 @@ func (r *SFController) DeployGitServer() bool {
 	}
 
 	// Define initContainer
-	initContainer := base.MkContainer("init-config", base.GitServerImage)
+	initContainer := base.MkContainer("init-config", base.GitServerImage())
 	initContainer.Command = []string{"/bin/bash", "/entry/pre-init.sh"}
 	initContainer.Env = []apiv1.EnvVar{
 		base.MkEnvVar("FQDN", r.cr.Spec.FQDN),