+++ controllers/libs/logging/logging.go
@@ -55,10 +55,7 @@ func SetupLogForwarding(serviceName string, forwarderSpec *v1.FluentBitForwarder
 }
 
 func CreateFluentBitSideCarContainer(serviceName string, extraLabels []FluentBitLabel, volumeMounts []apiv1.VolumeMount, debug bool) apiv1.Container {
-	var img = base.FluentBitImage
-	if debug {
-		img += "-debug"
-	}
+	var img = base.FluentBitImage(debug)
 	container := base.MkContainer("fluentbit", img)
 	container.Env = CreateForwarderEnvVars(serviceName, extraLabels)
 	ports := []apiv1.ContainerPort{