+++ controllers/nodepool.go
@@ -530,7 +530,7 @@ func (r *SFController) DeployNodepoolBuilder(statsdExporterVolume apiv1.Volume, 
 		"ssh_config":                 utils.Checksum([]byte(builderSSHConfig)),
 		"buildlogs_httpd_config":     utils.Checksum([]byte(httpdBuildLogsDirConfig)),
 		"statsd_mapping":             utils.Checksum([]byte(nodepoolStatsdMappingConfig)),
-		"image":                      base.NodepoolBuilderImage,
+		"image":                      base.NodepoolBuilderImage(),
 		"nodepool-providers-secrets": getSecretsVersion(providersSecrets, providerSecretsExists),
 		"serial":                     "14",
 		"corporate-ca-certs-version": getCMVersion(corporateCM, corporateCMExists),
@@ -540,7 +540,7 @@ func (r *SFController) DeployNodepoolBuilder(statsdExporterVolume apiv1.Volume, 
 		annotations["config-repo-info-hash"] = r.cr.Spec.ConfigRepositoryLocation.BaseURL + r.cr.Spec.ConfigRepositoryLocation.Name
 	}
 
-	initContainer := base.MkContainer("nodepool-builder-init", base.NodepoolBuilderImage)
+	initContainer := base.MkContainer("nodepool-builder-init", base.NodepoolBuilderImage())
 
 	initContainer.Command = []string{"/usr/local/bin/init-container.sh"}
 	initContainer.Env = append(r.getNodepoolConfigEnvs(),
@@ -571,7 +571,7 @@ func (r *SFController) DeployNodepoolBuilder(statsdExporterVolume apiv1.Volume, 
 	}
 
 	nb := r.mkStatefulSet(
-		BuilderIdent, base.NodepoolBuilderImage, r.getStorageConfOrDefault(r.cr.Spec.Nodepool.Builder.Storage),
+		BuilderIdent, base.NodepoolBuilderImage(), r.getStorageConfOrDefault(r.cr.Spec.Nodepool.Builder.Storage),
 		apiv1.ReadWriteOnce)
 
 	nb.Spec.Template.Spec.InitContainers = []apiv1.Container{initContainer}
@@ -602,7 +602,7 @@ func (r *SFController) DeployNodepoolBuilder(statsdExporterVolume apiv1.Volume, 
 	nb.Spec.Template.Spec.Containers = append(nb.Spec.Template.Spec.Containers, diskUsageExporter)
 
 	// Append image build logs HTTPD sidecar
-	buildLogsContainer := base.MkContainer("build-logs-httpd", base.HTTPDImage)
+	buildLogsContainer := base.MkContainer("build-logs-httpd", base.HTTPDImage())
 	buildLogsContainer.VolumeMounts = []apiv1.VolumeMount{
 		{
 			Name:      BuilderIdent,
@@ -714,7 +714,7 @@ func (r *SFController) DeployNodepoolLauncher(statsdExporterVolume apiv1.Volume,
 		"statsd_mapping":        utils.Checksum([]byte(nodepoolStatsdMappingConfig)),
 		"serial":                "8",
 		// When the Secret ResourceVersion field change (when edited) we force a nodepool-launcher restart
-		"image":                      base.NodepoolLauncherImage,
+		"image":                      base.NodepoolLauncherImage(),
 		"nodepool-providers-secrets": getSecretsVersion(providersSecrets, providerSecretsExists),
 		"corporate-ca-certs-version": getCMVersion(corporateCM, corporateCMExists),
 	}
@@ -723,7 +723,7 @@ func (r *SFController) DeployNodepoolLauncher(statsdExporterVolume apiv1.Volume,
 		annotations["config-repo-info-hash"] = r.cr.Spec.ConfigRepositoryLocation.BaseURL + r.cr.Spec.ConfigRepositoryLocation.Name
 	}
 
-	initContainer := base.MkContainer("nodepool-launcher-init", base.NodepoolLauncherImage)
+	initContainer := base.MkContainer("nodepool-launcher-init", base.NodepoolLauncherImage())
 
 	initContainer.Command = []string{"/usr/local/bin/init-container.sh"}
 	initContainer.Env = r.getNodepoolConfigEnvs()
@@ -756,7 +756,7 @@ func (r *SFController) DeployNodepoolLauncher(statsdExporterVolume apiv1.Volume,
 
 	nl := base.MkDeployment("nodepool-launcher", r.ns, "")
 
-	container := base.MkContainer("launcher", base.NodepoolLauncherImage)
+	container := base.MkContainer("launcher", base.NodepoolLauncherImage())
 	container.VolumeMounts = volumeMounts
 	container.Command = []string{
 		"/usr/local/bin/dumb-init", "--", "bash", "-c",